Grow the spoke gradient image to fit the nav_area (#1035772).
The nav_area might be taller than we expect on certain languages, like Kannada.
Because we were shrinking the nav_area to fit the size of the image, widgets
in the nav_area looked like they were spilling over.  Instead, we should force
the image to fit in however big GTK decides to make the nav_area.