Fix closest mirror showing up when it shouldn't (#1031663)
We are not shipping the fastest mirror plugin in RHEL7,
so the closest mirror source option should not be shown.

We do this by waiting for the payload processing thread
to finish - the payload detects the fastest mirror plugin
is not installed and the closest mirror option is removed
from the menu.

If the user tries to access the source spoke before the
source & payload processing is done, he can either press
r to try again or c to return to the hub and interact with
other spokes.