A class for scheduling Gtk actions and running them all at once
This could simplify some code and make it more granular in a way of what's run
in the main thread and what's run in non-main threads. The code can schedule
actions without waiting for each of them to be processed by the main loop and
running them all at once in the end.

There's no case we could use this class in our code now, but it may be useful
for addons and in the future.