Various changes to handling of filesystem label setting (#1038590)
Related: rhbz#1038590

1) Change validate so that it treats an empty label as accepting the default
if the filesystem is not able to assign labels.
2) If the value of label is None, meaning accept the default for the
filesystem, change it to "". This makes coping with changes in this field
much easier and retains the previous behavior pretty well.
3) Check the label if either the label or the filesystem has changed as the
label's validity is dependent on both these things.
4) Make _labelEntry sensitivity equivalent to non-existance of the format.
This means that sometimes the _labelEntry field will be sensitive even if
the currently chosen filesystem does not allow labeling, that is no worse
than the problems that arise from changing the sensitivity too sensitively.
5) Don't change _labelEntry sensitivity in a couple of places where it was
previously changing.
6) Don't use current _labelEntry sensitivity to infer the validity of the text
in _labelEntry field.

Toggling the label field sensitivity on and off could cause a lot of problems.
So, when populating the right side label field is set to sensitive if the
filesystem does not exist and is never changed anywhere else.

The user may only choose not to reformat if the filesystem already exists.
In that case, though, the user can not set the label. So, if the user has set
a label, then formatting will occur because the filesystem does not exist.
Therefore, no need to check if reformat is True when validating label.

Signed-off-by: mulhern <amulhern@redhat.com>