Be more liberal in what is accepted as a size unit.
Not all size specifications will end with a letter. For example: the
translation of "B" for byte in Telugu ends with a vowel sign,
categorized in Unicode as a non-spacing mark. Instead check whether the
size string ends with a [0-9]: if so, the unit is missing.

(cherry picked from commit a9e107499f1c99b11b6ab4780d5191e031367df2)

Related: rhbz#1075561