Don't add redundant grub installs if stage1 is not a disk (#1070283)
The GRUB bootloader handler attempts to install redundant stage1
bootloaders if stage2 is on a RAID. However, this assumed that if stage1
was not a disk that it was to be treated the same as stage2, which is
not the case when prepboot or biosboot partitions are used. Instead of
attempting to add redundant installs for the non-disk cases, just let
RAID handle the parts that were configured for redundancy and warn the
user that their bootloader will not survive a disk failure.

Resolves: rhbz#1070283

Signed-off-by: Brian C. Lane <bcl@redhat.com>