Fix problems going into custom partitioning with the new work flow.
(1) If the user's checked both the reclaim button and the custom button,
do not display the reclaim dialog.  That's what custom partitioning is
for.

(2) If the user's checked both the encryption button and the custom button,
do not display the passphrase dialog before custom partitioning.  It gets
displayed afterwards.

Related: rhbz#1014671