Fix tb due to non-existant disk attr. (#1054746, #1056019)
Adding a specialized device to the disk store was occurring twice
due to commit 6977f898cba63d0f215ca5e2be7c96bd451e8fc1, which is
unnecessary. The tb was occurring specifically because the disk
store addition attempted to access an attr not present on iscsi
or fcoe devices.

Resolves: rhbz#1054746
Resolves: rhbz#1056019