Set device.format.label field close to where we read it (#1056139)
Resolves: rhbz#1056139

Previously, somewhere between where this was set and where it was first read
in _populate_right_side it was occasionally being reset to None. This
should have been impossible as the code between is very simple and it should
be impossible for it to have this side-effect. Nonetheless, it was observed
to happen.

Signed-off-by: mulhern <amulhern@redhat.com>