Always reject label if the format exists (#1038590)
Related: rhbz#1038590

Previously the label was allowed if it was empty and the fs wasn't labeling,
because in that case "" stood for the default. This happened regardless of
whether or not the format existed. But if it exists, then it has a label
and we're not going to reformat, so accepting the default is meaningless.

No other semantic change, but some rearrangement of conditionals, for clarity.

Signed-off-by: mulhern <amulhern@redhat.com>