Use validate_label to check whether label should be updated (#1038590)
Related: rhbz#1038590

Some things can be updated even if no reformatting is going to be done,
because they represent the user just changing their mind from some previous
state that they had decided on, not actually changing the state of a device.
One of these changes is relabeling a filesystem if the filesystem hasn't
been created yet.

But, we should use validate_label to check if the label is OK in that
situation so that we have last chance to detect formatting issues.
validate_label(...) == LABEL_OK will be False if not device.format.exists is
False. But will also be False in a few other cases having to do w/ formats.

If neither the format nor the label has changed in between the previous check
and now, then the new line is unnecessary. But there's 300 lines in
between, so not taking any chances.

Still checking whether the label has changed, because do not want to
log if it hasn't.

Still checking whether the label field exists, validate_label may return True
if the label field does not exist, in the case where label itself is
the empty string.

Signed-off-by: mulhern <amulhern@redhat.com>