Keep selection when switching environments
When user selects an addon for an environment,
keep it selected if user switches to another
environment that also has the addon available.

Also if user deselects a default addon for an
environment, it will also be deselected when
switching to any other environment that has
the addon selected by default.

Related: rhbz#1043655