Put Xorg on tty6 in accordance with Ancient Anaconda Tradition (#980062)
Previously Xorg and logind were fighting over tty6. logind got there
first, so it won the race and Xorg ended up on tty7. commit 0a302ce moved Xorg to vt7 so we could avoid the race.

More recently, lorax commit dff1542 reconfigures logind to reserve tty2
instead of tty6. So now tty6 is free again.

So until we're ready to move Xorg to tty1 and start it up like a normal
system does, we might as well leave it where it was before.

This reverts commit 0a302ce1d00929e0f835925454caa63f24c10288.

Conflicts:
    anaconda