Put Xorg on tty6 in accordance with Ancient Anaconda Tradition (#980062)
Previously Xorg and logind were fighting over tty6. logind got there
first, so it won the race and Xorg ended up on tty7. So we moved Xorg to
vt7 so we could avoid the race entirely.

A recent lorax fix reconfigures logind to reserve tty2 instead of tty6.
So now tty6 is free again.

So until we're ready to move Xorg to tty1 and start it up like a normal
system does, we might as well move it back to it where it was before.