Display custom part warnings/errors on the spoke itself (#975840).
This works similarly to how the password spokes do:  If the user does
something incorrect, the info bar displays a message and the user can
click on it to see the problem.  If they really want to, they can click
done again to go back to the hub.

Note that in the case of errors, the user will still have to do something
about them even if they do leave the custom part spoke.  But, we don't
want to trap them.