Change the reclaim space button rules (#980496)
Now that the reclaim dialog can be entered with pre-existing free space
the button state needs to be updated with the initial amount of free
space. And its sensitivity should depend on whether or not there is
enough free space, not on delete actions, which may be 0 if you are only
shrinking a partition.

Change _recursiveRemove to check for protected children and only delete
the unprotected ones. If there are none, delete the whole device.

Related: rhbz#980496