Change the string used to test for serial console (#1054951)
The string used to test for serial console is a literal ("ttyS").  To support
serial consoles other than those starting with "ttyS", such as ttyAMA0 for ARM,
create a list of string values that may be used.  Add a property to test if
the console device starts with any of these strings.

Note: 64-bit ARM (aarch64) hardware may use either ttyS0 or ttyAMA0 for the
serial console.  The Foundation Model simulator currently uses ttyAMA0.

Resolves: rhbz#1054951

Signed-off-by: David A. Marlin <dmarlin@redhat.com>
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>