Add option help text for --image and --dirinstall flags (#1056791)
Resolves: rhbz#1056791

The help text is now in a separate date file anaconda_options.txt, as
putting it in-line would be distracting.
It is in a text file with very simple formatting requirements because the
more fancy data formats like configurations or JSON are kind of awkward to
edit by hand in the way that you would like to edit help text.

Signed-off-by: mulhern <amulhern@redhat.com>