If a user has been created, don't allow entering the user spoke (#1058564).
However, we still show the spoke in an insensitive state since removing all
spokes from the progress hub might look pretty weird.  This also redoes the
root password spoke fix from #1041405 to make it insensitive instead of
hidden too.

I've also changed the sort order of the TUI spokes so that the user and root
password are always sorted to the end.  This way if they do not appear, the
other spokes do not get numbered differently.