If a user has been created, don't allow entering the user spoke (#1058564).
However, we still show the spoke in an insensitive state since removing all
spokes from the progress hub might look pretty weird.  This also redoes the
root password spoke fix from #1041405 to make it insensitive instead of
hidden too.

Note:  Nothing should be required for the text UI here.  The password and user
spokes get sorted to the end of the list always.  This will require changing on
rhel7-branch.