Update bumpver to allow Related bugs
Related bugs that are against a different component will have mismatched
Fixed-in-version strings, which is to be expected. So if it is Related,
valid and has an ack override the bad flag and allow it.

(cherry picked from commit 76ee71bb3a4f285f6630da6e484d897cee4df8c2)