Move save_netinfo into a hook (#1048231)
There were a couple of issues here - echoing a command into a hookfile
isn't the right way to setup a hook anymore, use inst_hook in
module-setup.sh for that. And the script was being run after
11-fetch-kickstart-net.sh because of its name. It is now installed as
00-anaconda-ifcfg.sh

This also changes new_dhcp_next_server to new_next_server, which is the
variable that is actually set by dhclient. This lets it find the nfs
server when it is different from the dhcp server.

Resolves: rhbz#1048231