Prevent kickstart parsing errors from ending up in boot options (#1060184)
If errorsAreFatal is set to False, pykickstart only outputs the error
message but does not interrupt parsing. This results in the kickstart
parsing error message being included in the boot options file that is generated
by the parse-kickstart Dracut module.

By setting errorsAreFatal to True, parsing is interrupted on error,
the error message is correctly forwarded to the system log and the boot
options include only output of correctly parsed kickstart commands.