Properly retry package downloads (#924860)
There are basically two cases where package download can fail:
- when YUM is populating the transaction
  (this actually just checks the file exists)
- when the transaction is started and packages are downloaded and
  installed

Previously Anaconda did not attempt to retry the first one at all
and made an unlimited number of attempts in the second one.

This has been changed and Anaconda now does 10 retries separated
by a pause that progressively grows from 0.5 to 256 seconds for
the final wait before giving up. With this mechanism it should
be possible to successfully install even with a very shaky connection.

For providing the progressive delay a new generator function has
been added to iutil.