driverdisk: Add dd_args_ks handling to driver-updates (#1035663)
This reworks how driver-updates works to look for files in /tmp/ instead
of expecting a cmdline arg (although that does still work). This allows
us to use a single systemd service and to turn the UI startup into a
common function in anaconda-lib.sh

If /tmp/DD-net/ is present it processes the drivers or iso's in the
directory that have been downloaded from a network source. Passing a
path to driver-updates also triggers this behavior.

If /tmp/dd_args_ks is present the existing OEMDRV devices are ignored
(they will have been processed eariler) and the partitions in
/tmp/dd_args_ks will be processed, as well as any OEMDRV devices that
may appear as a result.

If neither /tmp/dd_args_ks nor /tmp/DD-net/ are present it will look for
OEMDRV labeled devices and process partitions from /tmp/dd_args if
present.

Resolves: rhbz#1035663