driverdisk: Use a single systemd service to start DD UI (#1035663)
This removes the second driver-updates service for network UI startup
and creates a start_driver_update function that is used by the 3 places
that driver updates need to be processed.

It also moves the driver-fetch-net to run after kickstart-fetch-net so
that kickstart driverdisk network sources can be processed in the same
way as the cmdline network sources.

Related: rhbz#1035663