kickstart user accounts should be locked by default (#1063554)
Commit 18a8f661635884e7861bc14f200455e470bebb0b changed createUser so
that an account could be created with an empty password (see bug
--password or --lock argument being created with no password, which is
not what was intended.

The impact was limited to local accounts, ssh blocks access to accounts
with empty passwords by default. But another local user could access one
of these accounts by using su.

This patch fixes user creation so that when accounts are created from a
kickstart and their password argument is blank (the default from
pykickstart is "") it will switch it to None so that createUser will
lock the passwordless account.

Resolves: rhbz#1063554