Move environmentAddons into packaging
The environment addon refresh really only needs to be executed when new
metadata has been fetched so move it into packaging and call it after
metadata download. It was also running very slow from the other thread,
moving it speeds it up considerably.

This also removes the need to call it from software.py when the source
has changed, since it will already be updated when the spoke is entered.