Use a single script to run the glade tests.
All of the glade/run_*.sh do pretty much the same thing; we don't need
four copies of it. Modified the pw_visibility and validity checks to
accept and ignore the translation arguments and added a script to run
all the glade checks on all the glade files.

This change means that the tests won't run (potentially) in parallel,
but that shouldn't make a huge difference.