Preserve ipv6.disable=1 on target system (#1040751)
In Olden Times anaconda had the "noipv6" boot argument, which disabled
ipv6 in the installer and on the target system. It did this by adding
"options ipv6 disable=1" to /etc/modprobe.d/ipv6.conf, so ipv6 would be
disabled before the network came up.

Now that ipv6 isn't modular, we can't do that. In an attempt to retain
backward-compatibility, "noipv6" now causes anaconda to disable ipv6 via
sysctl (i.e. in /etc/sysctl.d/anaconda.conf) - *iff* all network devices
have ipv6 disabled (by 'network --noipv6' kickstart commands or the
'ipv6.disable=1' boot argument).

This can be a bit confusing:
 * If you netboot with 'noipv6' but no kickstart, the system gets ipv6.
 * If you netboot with 'noipv6 ipv6.disable=1', the system does ipv6 for
   a moment, then disables it via sysctl.
 * Neither boot argument is preserved, so it's unclear to users
   how/whether the system has ipv6 disabled.

This patch adds 'ipv6.disable' to global_preserve_args. This way,
booting with "ipv6.disable=1" will:
  a) disable ipv6 in the installer, and
  b) disable ipv6 on the target system before the network comes up.
This is a lot closer to the original behavior, and a lot simpler for us
as well.