let systemd decide when to start anaconda-sshd
This just moves the logic for starting anaconda-sshd into the systemd
unit file(s), so there's less stuff for us to maintain.

Note the '/sys/hypervisor/s390' test to see if we're on s390 (and
therefore our console is probably bunk). Later we can change that to
ConditionArchitecture, which is slightly cleaner; Ideally we'd want to
actually check the console capabilities but that's way trickier than
you'd think.