Don't traceback when no size is given in kickstart (#1067707).
There are various circumstances where a size may not be given - recommended swap
size, grow, and the like.  We need to ignore these situations and let the
methods we call deal with the default value.  This bug was caused by
ea61ee97601b3aa3f49c2163d0b86c48ef6a3e31.  I've added a test case for it.