Remove a bunch of unused includes and tests for headers
Most of isys.c is no longer relevant, and a fair amount of the autoconf
tests are for things that used to be in isys or loader or other code
that is no longer with us. Also anaconda barely links to anything
through C, so there's no need to pull in a bunch of CFLAGS and LDFLAGS
for various packages. Moved the one actual bit of information from all
that--the required Gtk version--to widgets, where the newer Gtk
functionality is actually used.