Don't use tmux for inst.noshell (#1058607)
If we're running with 'inst.noshell', then the system administrator
wants this to be a hands-off install. And tmux is very hands-on, and it
lets you create new shells if you want. So we should probably disable
tmux when running with inst.noshell.

This patch moves the logic for deciding when to use
anaconda-direct.service or anaconda.service out of anaconda-generator
and into the units themselves. The behavior should mostly be the same as
before, i.e.:

anaconda-direct on s390, otherwise anaconda+tmux

But now we want to use anaconda-direct instead of tmux when we have
inst.noshell, so it's:

anaconda-direct on s390 or with inst.noshell, otherwise anaconda+tmux