text install -> text system (#1021963)
Doing as RHEL once did: if you're installing in text mode, the resulting
system should default to text mode.

So, after all the memory checks etc. happen, if you're still not in
graphical mode, you're in text mode. So we add 'skipx' to your kickstart
data and go from there.