Write 'text'/'cmdline' in anaconda-ks.cfg in text/cmdline mode
If the install is happening in text/cmdline mode, we should end up with
'text' or 'cmdline' in the resulting anaconda-ks.cfg.

So: once we have determined anaconda.displayMode, set the appropriate
ksdata.displaymode.displayMode so that we'll write it to the kickstart
later.

(Note that we do this after setupDisplay because setupDisplay can change
displayMode - e.g. by forcing text-mode on low-memory systems or by
asking the user if they want to use VNC instead.)