Let Gtk pick the size for the isoChooserDialog (#973376)
If it's not big enough now, it's not our problem.

Also converted the dialog from a GtkDialog containing a
GtkFileChooserWidget into just a GtkFileChooserDialog, which is
basically the same thing but one less layer of objects that we have to
define for ourselves.

Unrelated things that glade changed but that seem ok enough to keep: a
<style> element was moved from the middle of an <object> element to the
end of it, and two instances of "icon-size" were changed to "icon_size".