Fix console for s390 and 'noshell' mode (#1070672)
So the way sshd installs work, we actually do need to run tmux on s390 -
when the user connects to the system, it'll attach them to the existing
tmux session.

Relatedly, we also actually need to use tmux when in 'noshell' mode,
because anaconda-direct.service doesn't allow any input from stdin, so
e.g. you can't answer the VNC question.

So: instead of disabling tmux for noshell or s390x, we should do the
following:

1) Start anaconda inside a detached tmux session
2) If "noshell", lock down tmux (close the shell, disallow new windows)
3) If s390x, run anaconda-direct.service like usual

This commit also silences an error message about the missing 'null' in
anaconda-direct.service.

Resolves: rhbz#1070672