driver-updates: refactor dd_scan
Three refactoring changes:

1) Make default value for skip_dds=set()
The docstring says the value will be a set(), so make it a set().

2) Set default state for dd_todo/dd_finished outside the if/else
This is just for clarity, mostly.

2) split the dd autoload loop into its own function, dd_load()
This is useful because we might want to do the dd_load() loop in more
than one place. Like, say, after each DVD, if we were loading drivers
off sequential DVDs.

Note that we're filtering the skip_dds out of dd_todo twice:

  * dd_scan() removes dd_finished (an alias for skip_dds) from dd_todo
  * dd_load() skips anything in skip_dds

This is redundant, and one of them can probably be removed.

Related: rhbz#1066784

Conflicts:
	dracut/driver-updates