driver-updates: add 'refresh' to selection_menu()
This allows the user to signal that they would like to refresh the
contents of the menu. You'll need a loop that wraps selection_menu() and
restarts the menu (with refreshed contents) whenever DoRefresh is
returned.

There are better ways to do this, but we don't have the time for that
right now.

Related: rhbz#1066784