RAID related changes for custom spoke.
- RAID objects instead of strings are now used everywhere. Previously,
the approach was a bit more hybrid, with strings floating around here
and there.

- Move get_supported_raid_levels to blivet.devicefactory. This method
is supposed to yield the raid levels that blivet supports, not necessarily
those that anaconda allows. The return type is changed to a set of RAID
objects during the move and the return values are somewhat modified.
* btrfs substitutes single for none
* RAID adds container and linear

- Currently RAID levels are only offered on the Manual Partitioning screen
if the user has selected RAID. So, the only options that are ever shown are
one that RAID allows.

- In CustomPartitioningSpoke._validate_mountpoint RAID related validation
tests have been generalized a bit.

- get_raid_level in custom.py replaces blivet.devicefactory.get_raid_level.
The return type is changed to a RAIDLevel object.

- In _save_right_side the container raid level validation has been
generalized a bit.

- CustomPartitioningSpoke._raid_level_visible has been generalized to
use raidLevelsSupported().

- CustomPartitiongSpoke._populate_raid has been simplified a bit,
and generalized a bit.

- In on_device_type_changed some general methods are used. Previously the
code sets btrfs's raid_level and passed it to _populate_raid(), but
_populate_raid() never displays any RAID choices for btrfs on manual
partitoning screen, so this code had no effect. Btrfs RAID choices are
shown on the "Configure Volume" screen, reached by pressing the Modify
button. The effectless code is just removed. This leaves a few unused
identifiers, which are also removed.

- In the "Configure Volume/Configure Volume Group" window, "single" is added
for btrfs.

- selectedRaidLevel() is changed so that it returns a RAIDLevel object or
None, rather than a string.

- Several methods are added to custom.py. They are small
and copiously commented.

- ContainerDialog._raid_level_visible can be simplified by using
containerRaidLevelsSupported(), one of the newly introduced methods.

- ContainerDialog._populate_raid() can also be simplified a bit.

- Use StorageDevice.raw_device where appropriate.

Incorporates some commits cherry-picked w/ changes from master branch:
c8538b0cfd726deffabfbcda837239af2b942014
aeff4ca3f7b7d813bb7080a9e10281f599a45802

Most of this commit was cherry-picked from
4d41d2cbd4342e1f1e8e853531526954da3b6829, with quite a number of changes.

Signed-off-by: mulhern <amulhern@redhat.com>

Related: rhbz#1075561