Add rescue kernels to the bootloader install list. (#1036349)
This adds a new property, rescueKernelList, to Payload.
kernelVersionList is unchanged, since most other places that use it
excluded rescue kernels for a reason.

Every bootloader but grub2 generates a config based on the list of
kernels provided by the Payload object, so every bootloader but grub2
was skipping the rescue kernels. This remedies that.