driver-updates: accept burned driver discs (#1073719)
driver-updates was written with the assumption that the interactive menu
would always involve choosing a device which contained multiple updates
images.

This is not really the case - that's not how it worked in RHEL6, for
example - so we need to handle the case where the selected device is
itself a driver update disc.

This patch just checks the chosen device - if it is, in fact, a driver
update disc, then we unmount it (so the later functions can re-mount
it) and return that device name.

(Leaving the mountpoint as None is valid, as our subsequent umount()
will simply ignore that.)

Also, for the sake of sanity, remove is_dd(). A previous version of this
patch tried to use it, and it's a) otherwise unused and b) doesn't work.

Bonus: remove a debugging "print" statement that was never supposed to
be there. MY BAD Y'ALL.

Resolves: rhbz#1073719