Fix environment group changes based on ListBox row activation
i.e., clicking somewhere in the environment row that isn't the radio
button. When we call button.set_active in on_environment_activated, two
button toggles are emitted: the one for the button becoming active,
which we block, and the one for the button becoming inactive, which was
causing us to reactivate the inactive row. Using "clicked" handlers
instead of "toggled" handlers was probably supposed to get around this,
but buttons going inactive receive a "clicked" signal as well, because
nothing means anything.

Switch the button signal handlers to connect to "toggled" instead of
"clicked" to ensure consistent behavior with the set_active calls,
use separate the toggled handlers for the environment radio buttons and
the addon check buttons, and ignore any radio button toggle that is
toggling to an inactive state. Toggle toggle toggle.