Wait for other threads to finish before sending ready (#1075103)
More threading fallout. In _initialize we set self._ready and send the
hub the spoke ready message. But some of the other threads may not be
quite done when the hub processes the message so it doesn't set the
Source spoke sensitive. And it stays that way until another spoke is
visited and _updateCompleteness is called on all the spokes.

This adds a loop at the end of the _initialize thread to make sure
self.ready will really be true when the hub gets the source spoke's
ready message.

Resolves: rhbz#1075103