Convert iter from filter model iter to backing store iter (#1074188)
Resolves: rhbz#1074188

Previously the backing store was being indexed by means of an iter that
was a correct index into the filter model. It was not a correct index
into the backing store, so the set of disks checked by the user
was not the same as the set of disks actually selected.

Now the index is converted, so the set of checked disks and the set of
selected disks should be the same.

Signed-off-by: mulhern <amulhern@redhat.com>