Append cmdline arg values in BootArgs (#1073130)
This creates a space separated list of the argument values when there is
more than one occurrence of a boot arg (eg. modprobe.blacklist can be
used multiple times). I did it this way instead of making a list
so that BootArgs users that don't have to check for list vs. string.

Related: rhbz#1073130