safe_dbus: Don't export DBus connection addresses as variables
This required a previous commit to do a workaround to prevent pylint
from attempting to spawn a bus while processing imports.  This patch
fixes it in a cleaner way by not defining variables at all for them.

For the system bus, add a helper to get a complete connection as an
API.

For the session one, since we only use it in one place, so let's just
internalize it.