Add support for tarfiles to liveimg kickstart command (#1080396)
With these changes you can now pass the url of a tarfile to liveimg and
it will install it to the target system. The same rules apply, the
correct utilities need to be included in the tar for the configuration
to work correctly.

The following extensions are recognized as tarfiles:

.tar .tbz .tgz .txz .tar.bz2 .tar.gz .tar.xz

Resolves: rhbz#1080396
(cherry picked from commit ccb9c7729234d681b85dca584a79170951ff0549)

Related: rhbz#1196721