Add support for tarfiles to liveimg kickstart command
With these changes you can now pass the url of a tarfile to liveimg and
it will install it to the target system. The same rules apply, the
correct utilities need to be included in the tar for the configuration
to work correctly.

The following extensions are recognized as tarfiles:

.tar .tbz .tgz .txz .tar.bz2 .tar.gz .tar.xz