Avoid the "unable to init server" message.
This message is only printed by the "broadway" backend, and the broadway
backend is only used because Gdk tries every backend if it can't figure
out what is going on, so just don't use that backend. The content of the
"Unable to init server" message changed recently which is why it started
getting past our false-positive filter again.