Use an alternative image if logo is missing
If the logo in upper left is missing GTK will start spamming
TT1, making it completely unusable. The image is specified in a CSS
and needs to be present on the given path before the GUI is started.
Therefore we need to check if the logo image is present and symlink
a replacement image to the logo path if the logo image is missing.

By default an Anaconda Fedora header image will be used that looks
quite OK, but might not be always present (comes from the fedora-logos package).
So there is a fallback to a GTK icon, which should be always available as we use
GTK icons for Anaconda UI elements.