Don't do yum lock logging when using updates.img
The commit d42a7caa turned on debug logging when using updates.img. However,
that also turns on yum lock logging which is kind of a flood for the
packaging.log and which prolongs anaconda-tb file generation a lot. We only want
to do yum lock logging when inst.debug is used.