mountExistingSystem raises an exception with dirty FS (#1080210)
The commit 3ca0ae5e155 removed the rc variable that had always been being set to
None because mountExistingSystem didn't return any value. However, the variable
was left in place in the next block testing it for being -1 or something else.

Put mountExistingSystem call in a try-except block and catch the DirstFSError
exception instead.