Remove stock labels and icons.
Stock labels are deprecated now. This is actually good, because the
stock labels defined keyboard accelerators that are opaque to us, which
is bad.

Diverged from Gtk's migration advice in the cases of gtk-yes and gtk-no,
since there was no advice, and with gtk-find and gtk-clear, since they
said to just switch to a named icon which is dumb.