Fix the way categories are handled in text mode.
This statically maintained list of spoke categories for text mode has
been bothering me for almost a year, so it's time to get rid of it.

This now more or less mirrors the GUI logic we have for handling spoke
categories. I wanted to originally make the category classes something
that could be used by both GUI and TUI to reduce duplication, but it is
not meant to be.

This also moves the collectCategoriesAndSpokes, collect_categories, and
collect_spokes functions into pyanaconda.ui.common since I could at
least consolidate that.