Add a new InputCheck status for silent failures.
Add a constant CHECK_SILENT as a return value for checks that should be
considered failed but that should not be displayed as error messages.
This can be used to, for example, to keep widgets insensitive for empty
input but not have empty trigger an error message.

Add a new generator to InputCheckHandler to differentiate between failed
checks and failed checks with error messages.