Don't overload selectorFromDevice function
The selectorFromDevice either creates a new selector from a device or updates an
existing one depending on the (not-)passed selector parameter. This makes its
usage quite confusing. Easy way to streamline that is to split the function into
two functions -- newSelectorFromDevice and updateSelectorFromDevice.

Note that I'm keeping the camelCase naming due to the surrounding code.