Add the beginnings of an automated GUI test suite.
This is enough to make sure all the basic stuff works on a standard livecd.
The documentation throughout should be enough to explain it all, but here's
a few quick comments:

* Everything in the inside/ directory runs inside a VM.

* Most things in the outside/ directory run outside of and control a VM.
outside/template.py is a notable exception to this.  It is a template file of
python code and gets run inside the VM as suite.py.

* runtest.sh runs all the outside/ tests against a livecd previously created
with make_livecd.sh.