Make on_updates_settings_clicked timed callback
Updating mountpoint settings often takes quite a lot of time when the installer
seems to be hanged. With the @timed_action decorator it first calls busyCursor(),
then runs and then calls unbusyCursor() providing better UX. With such short
delay and threshold parameters, it means the callback is run basically instantly
just with a time for Gtk to switch the cursor.