Add input validation to the source spoke.
Validate the network URL if using a network. Convert the duplicate repo
name and invalid repo name checks into InputChecks. Add checks for
invalid addon URLs and proxy URLs. Prevent the user from leaving the
spoke if there is invalid data instead of simply discarding it. For the
repo field checks, hitting Done with failed checks will change the
selection in repoTreeView to the repo with invalid data and focus the
problematic input box.