Fix unloading modules in driver-updates (#1085099)
An unintended consequence of commit de2d663 is that driver-updates won't
unload *any* modules at all: /tmp/dd_modules (the list of "early"
modules that shouldn't be removed) is now created *after* we load all
modules.)

This patch makes driver-updates attempt to remove any driver that's in
the update disk(s) it loads.

First we make dd_extract() return a list of the firmware/module files
that were in the driver disk it extracted.

Next process_dd() builds a list of those for each extracted driver and
passes that to reload_modules().

reload_modules() maps those filenames to module names, and adds them to
the list of modules to be unloaded.

NOTE: Probably the correct behavior (and the way it worked in RHEL6)
would be to revert the change to /tmp/dd_modules and fix that bug in a
different way (e.g. by tracking driver disks by UUID instead of device
name, so they don't get loaded twice if the device name changes).

That requires a bunch of refactoring work; this is a stopgap solution.