Add viewports for the ListBoxes in the software spoke.
We had been defining the GtkListBox objects as direct children of the
GtkScrolledWindows, since an automatic viewport was added this way and
using an explict viewport broke the keyboard bindings. However, glade
seriously did not like this, and the bug in gtk appears to be fixed, so
adding explicit viewports.