iutil: Introduce getSysroot()/getTargetPhysicalRoot(), use instead of ROOT_PATH
For OSTree, the OS lives in a chroot, thus creating a distinction
between the "physical" root and the "deployment" root.  The physical
root as mounted by anaconda remains as /mnt/sysimage of course, but
*nearly* everything needs to reference the deployment, as that's where
the binaries, config files, etc live.

The deployment root is dynamic - we only know it after we do an
initial installation what the path will be.  This dynamic nature means
that the API is best as a function call.

Resolves: rhbz#1113535
Port of rpmostreepayload patches from master
commit 1b260c0c3ca70b53d44b543babe1a5ab06fb5fb5