install: Handle distinct physical root/sysroot
Recent changes introduced the ability to have distinct physical root
and sysroot.  In this case, because we run the bootloader binaries
inside the target, we need to ensure that e.g. /boot is mounted inside
that target root.

Note in this case, we also use a new Blivet API to update its
knowledge of the system root.

This patch also updates the storage write logic to note if
ksdata.ostreesetup.osname is set, because the storage will have to be
written after the packaging.

Resolves: rhbz#1113535
Port of rpmostreepayload patches from master
commit 0bbc9adf41b33062bbbfe478b3373a3404de21aa