Don't crash on anaconda-yum output containing multiple colons (#1092473)
The only reason this did not crash previously is that lines returned by the
anaconda-yum script did not contain more than one colon.
But once a new message with two colons has been added as part of the package download
retry feature, the split started to return 3 items, triggering the crash.