anaconda.service: Set GIO_USE_VFS=local
While a lot of the glib/gio stack is nice, it does have some desktop-y
defaults, such as this one to use the GVFS backend for GFile by
default.  This causes seemingly innocent looking API such as
g_file_delete() to do synchronous DBus calls on a session bus, which
we don't want since we're 1) running as root 2) outside of a
traditional session.