main: Set flags.extlinux if extlinux is used in interactive-defaults.ks
For making a custom Project Atomic boot ISO, I tried adding
"bootloader --extlinux" to /usr/share/anaconda/interactive-defaults.ks.

Confusingly this doesn't work because extlinux is a flag, and not just
a kickstart variable; the flag is set by the dracut KS parser. (Does
extlinux really need to be a flag?)

This small patch just ensures the flag is set if the KS enables
extlinux.

Note: I left a log.info in there since it seems useful to log when
we're parsing a KS file, but feel free to remove it.

Resolves: rhbz#1113535
Port of rpmostreepayload patches from master
commit f4921edf42d57523e8b42ba21d79e78a47d88e8b